Fix apidoc building
The apidoc building was failing in make distcheck because it was being
built using the standard 'html' target.  That target needs the output
to not be rebuilt each time make install-html is called.  Which is not
easily possible for us.  So for now, the apidoc is being generated
using the html-doc target.

	* doc/Makefile.am: Trigger the building frm the html-doc target.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>