Show the pretty representation of deleted variables in the diff output
While working on the 'abicompat' tool, it appeared that only the names
of deleted variables where printed in the diff report; what we want is
the full pretty representation of the variables, that is, we want
their type too.

This patch changes the reporting code of the corpus_diff to show the
full pretty representation of deleted variables.

	* src/abg-comparison.cc (corpus_diff::report): Show the pretty
	representation of deleted variables in all cases.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>