Better handle pointer-to-void in various places of the pipeline
This is the last patch that should fix the issue
https://sourceware.org/bugzilla/show_bug.cgi?id=17655

A pointer to void can be represented as an instance of
abigail::ir::pointer_type_def which has a NULL pointed-to attribute.

Unfortunately, there are various places in the code that assume that
abigail::ir::pointer_type_def::get_pointed_to_type() returns a
non-null value.  This patch is an attempt at fixing those spots.

	* src/abg-comparison.cc (pointer_diff::report): Handle the case of
	pointer to void.
	* src/abg-dwarf-reader.cc (build_pointer_type_def): Assert that
	the underlying pointer is non-null.
	* src/abg-ir.cc (pointer_type_def::pointer_type_def): Handle the
	case of pointer to void.
	(equals): Likewise in the overload for const pointer_type_def.
	(pointer_type_def::get_qualified_name): Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>