Tighten the condition for creating a cloned function from DWARF
So, apparently, the DWARF reader can be too eager to clone functions
whose DIE have a DW_AT_abstract_origin.  It seems to be that there are
cases where the second DIE (the one that has the DW_AT_abstract_origin
attribute) has the same linkage name than the first one.  In that
case, no cloning should happen.  And this should fix
https://sourceware.org/bugzilla/show_bug.cgi?id=17861.

	* src/abg-dwarf-reader.cc (build_ir_node): Re-indent.  Also,
	consider that when a DIE C refers to a DIE A via the
	DW_abstract_origin attribute, C represents a clone of A, only if C
	and A have *different* linkage names.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>