Speedup some diff::has_changes() implementations
Some implementations of diff::has_changes() are high on performance
profiles. This patch make them faster by trying to detect as quickly
as possible cases where the diff node actually has changes.  The
longest past is to detect when the diff node does *not* have changes.
We'll deal the latter case later.

	* src/abg-comparison.cc ({distinct_diff, var_diff,
	class_diff}::has_changes): Use the hash value of the diff subjects
	to detect quickly if they differ.  If they don't, then go the slow
	path of comparing the types.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>