Share private data of class_diff nodes
Profiling has shown that there can be a *lot* of class_diff nodes that
actually represent the same diffs.  In other words, these instances of
class_diff that is they are in the same equivalence class.  In that
case the private data of these class_diff node consume a lot of
redundant memory.  This patch is an optimization that leverages that
insight.  It shares the private data of the class_diff nodes that are
in the same equivalence class.

This makes the memory consumption of abidiff on the Xorg binaries of
RHEL 6 and 7 drop from more than 6GB to less than 370MB; execution
time drops from 15 to 7 minutes.

	* src/abg-comparison.cc (class_diff::class_diff): Do not
	initialize the private data of class_diff here.
	(compute_diff): In the overload for class_diff, initialize the
	private data of the new instance of class_diff to the private data
	of its canonical instance.
	(redundancy_marking_visitor::visit_begin): If a node is marked
	redundant, do not dare visit its children.  In cases of classes
	that have members that reference themselves, this prevents us from
	wrongly marking some of the data member changes as being
	redundant.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>