Avoid creating multiple versions of certain composite types
Sometimes during the reading of debug info, when creating a given
composite T, calling build_ir_node_from_die() to get the IR node for
their sub-type might lead to the creation of T, while we are in the
function that is supposed to create it.  In that case, just return the
type that has then be created from underneath us, rather than creating
a new one.

	* src/abg-dwarf-reader.cc (build_qualified_type)
	(build_pointer_type_def, build_reference_type, build_array_type)
	(build_typedef_type): If the composite type we are about to create
	was already created, just return the one that exists already.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>