Stick qualified, pointer, reference and array types into the global scope
In the debug info of least up to GCC 4.4.x, pointer, reference, array and
qualified types were in the global scope.  In GCC 4.8.x they can
belong to the scope of their sub-type.  The comparison code of
libabigail can thus (wrongly) consider that a qualified type described
by GCC 4.4.x debug info is different from the debug info of the *same*
qualified type emitted by GCC 4.8.x just because their scopes are
different.  The scope of qualified, pointer and reference types should
not matter anyway.  So this patch makes these composite types belong
to the global scope, irrespective of where they appear in the debug
info.  I have seen this when comparing libstdc++ from RHEL 6.5 and 7.
This is visible now that we have type canonicalizing.

	* src/abg-dwarf-reader.cc (build_class_type_and_add_to_ir): Do not
	consider qualified, pointer, reference and array types as member
	types.  Only typedef, class and enum types are.
	(build_ir_node_from_die): Stick base, pointer, reference,
	qualified and array types into the global scope.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>