Remove overly eager assert in distinct_diff::report
Since we now have proper diff node filtering capabilities, it appears
that there can be distinct diff node that is deemed to be reported
even though it's child node is not; this happens when the distinct
diff node does carry local changes.  So remove the assert that was
saying otherwise and enjoy one less abort.

	* src/abg-comparison.cc (distinct_diff::report): Remove over-eager
	assert.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>